// src/app/shared/model/scheduled-flight.ts
export class ScheduledFlight {
    scheduleFlightId: number;
    flightId: number;
    arrival: Date;
    sourceAirport: string;
    departure: string;
    destination: string;
    availableSeats: number;
    scheduleId: number;
    id: number;
    dateCreate: Date;
    updateDate: Date;
    departureTime: string;
    arrivalTime: string;

}